// web/widget/gql/schema.ts
export type Maybe<T> = T | null | undefined
// ...
export enum SlotStatusT {
  Enabled = 'ENABLED',
  Disabled = 'DISABLED',
  Unlisted = 'UNLISTED',
  Deleted = 'DELETED',
}
// ...
export type SlotByIdQuery = {
  slot?: Maybe<{
    id: string
    name: string
    description?: Maybe<string>
    notes?: Maybe<string>
    pricePerHour: any
    photoUrl?: Maybe<string>
    status: SlotStatusT
    ownerId: string
    address?: Maybe<{ [key: string]: unknown }>
    location: { longitude: number; latitude: number }
    slotAmenitiesList: Array<{
      id: string
      amenity?: Maybe<{
        id: string
        name: string
        description?: Maybe<string>
        slug?: Maybe<string>
      }>
    }>
    slotAvailabilitiesList: Array<{ dayOfWeek: number; startHour: string; endHour: string }>
  }>
}